Add API requests to login/register
Had to mess with /register a bit to make it work for me, authenticate() with custom callback apparently does not call req.login(), and changed bodyParser to take json. Also changed /register to return status numbers. Finally, added axios as a dependency instead of using the fetch API, in case we end up wanting to use cookies?